Give a better file name, flip lat/lon
